% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_forcing_args.R
\name{add_forcing_args}
\alias{add_forcing_args}
\title{Add Default Argument to Forcing Function Calls}
\usage{
add_forcing_args(expr, forc, prefix = "forc_")
}
\arguments{
\item{expr}{character vector of model expressions (processes or stoichiometry)}

\item{forc}{character vector or forcing functions for which an additional
argument is to be inserted}

\item{prefix}{prefix preceeding the generated argument name}
}
\value{
vector of modified model expressions
}
\description{
This function edits calls to time dependent forcing functions, e.g. 'foo(time)'
  to contain an additional parameter for defaults, e.g. 'foo(time, const_foo)',
  so that forcing functions can be handled as constant parameters, e.g. for model
  coupling in 'stop-and-go' mode.
}

