% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.stoi.R
\name{as.xstoi}
\alias{as.stoi}
\alias{as.xstoi}
\title{Convert Stoichiometry Tables}
\usage{
as.xstoi(x)

as.stoi(x)
}
\arguments{
\item{x}{crosstable resp. database table}
}
\value{
object of class stoi resp. xstoi
}
\description{
Convert stoichiometry tables from data base to crosstable format and and
Vice-Versa.
}

