% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_signals.R
\name{create_signals}
\alias{create_signals}
\title{Create \code{approxfun} Signals from List of Dataframes}
\usage{
create_signals(L)
}
\arguments{
\item{L}{list of data frames (currently with two columns only)}

\item{...}{further arguments passed to \code{\link{approxfun}}

//@return list of \code{\link{approxfun}}-functions (not yet implemented)}
}
\description{
Create a list of approxfun signals from a list of data frames with 2 columns
where the first column is taken as \code{x} and the second as \code{y}
}

