% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRodeoExt.R
\name{createRodeoExt}
\alias{createRodeoExt}
\title{Create an Extended 'rodeo' Model Object}
\usage{
createRodeoExt(folder, modfile, forcings, sources, modname)
}
\arguments{
\item{folder}{the folder where the model tables are located in}

\item{modfile}{the file with the model tables}

\item{forcings}{table with forcing functions}

\item{sources}{additional Fortran source files}

\item{modname}{name of the model}
}
\value{
an object of class \code{rodeoExt}
}
\description{
Create a model object from tabular data and save additional information.
}

