% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{Load Binary 'rodeo' Model From File}
\usage{
load_model(file)
}
\arguments{
\item{file}{full file name of the .rda object}
}
\description{
Load a model from class 'rodeo' or a derived class from a binary .rda file.
This requires that the corresponding class definition is already available
in the work space.
}

