% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.deSolve.R
\name{rbind.deSolve}
\alias{rbind.deSolve}
\title{Combine two deSolve Output Objects}
\usage{
rbind.deSolve(out1, out2)
}
\arguments{
\item{out1}{first deSolve object}

\item{out2}{second deSolve object}
}
\value{
combined ode object
}
\description{
This function is a stub .... ToDo: error checking, "diagnostics"-method
}

