% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_folder.R
\name{read_data_folder}
\alias{read_data_folder}
\title{Read All Data Files From a Folder}
\usage{
read_data_folder(path, pattern, header = TRUE, sep = "\\t", ...)
}
\arguments{
\item{path}{a character vector of full path names}

\item{pattern}{regular expression matching the file names}

\item{header}{boolean, if the first row of the data fioles is a}

\item{sep}{separator}

\item{...}{further arguments passed to \code{\link{read.table}}}
}
\value{
list with all data
}
\description{
Read all textfiles from a specified folder and return list with all data.
}

