% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlmodel.R
\name{read_xlmodel}
\alias{read_xlmodel}
\title{Read Model from Excel File and Return Text Representation}
\usage{
read_xlmodel(xlfile, sheets = c("vars", "pars", "funs", "pros", "stoi"))
}
\arguments{
\item{xlfile}{valid excel file with rodeo model description}

\item{sheets}{names of the spreadsheets imported}
}
\value{
list of data frames
}
\description{
Read Model from Excel File and Return Text Representation
}

