% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodeo_tables.R
\name{rodeo_tables}
\alias{rodeo_tables}
\title{Read 'rodeo' Model from Excel or CSV Files and Return Text Representation}
\usage{
rodeo_tables(folder, sheets = c("vars", "pars", "funs", "pros", "stoi"),
  type = c("Excel", "csv2"), ext = NA, ...)
}
\arguments{
\item{folder}{valid folder name or excel file with rodeo model description}

\item{sheets}{names of the spreadsheets imported}

\item{type}{type of the data format}

\item{ext}{file extension of text and csv files}

\item{...}{other arguments passed to the file read function, e.g. to
\code{\link{read_excel}} or \code{\link{read.csv2}}}
}
\value{
list of data frames
}
\description{
Read 'rodeo' Model from Excel or CSV Files and Return Text Representation
}
\details{
The tables must follow the rodeo specification and contain valid column headers.
}

