% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedpoly.R
\name{stackedpoly}
\alias{stackedpoly}
\title{Plot Stacked and Filled Polygons}
\usage{
stackedpoly(x, columns, density = NA, angle = 45, border = NA, col = NA,
  xlab = "x values", ylab = "y values", ...)
}
\arguments{
\item{x}{object of class \code{data.frame}, the first column is used as x-axis.}

\item{columns}{vector of column names in appropriate order that are plotted}

\item{density}{the density of shading lines, in lines per inch.
The default value of NULL means that no shading lines are drawn.
A zero value of density means no shading nor filling whereas negative
values and NA suppress shading (and so allow color filling).}

\item{angle}{the slope of shading lines, given as an angle in degrees
(counter-clockwise).}

\item{border}{border of the polygon}

\item{col}{the color for filling the polygon. The default, NA, is to
leave polygons unfilled, unless density is specified. (For back-compatibility,
NULL is equivalent to NA.) If density is specified with a positive value
this gives the color of the shading lines.}

\item{xlab}{label of the x axis}

\item{ylab}{label of the y axis}

\item{...}{further arguments passed to the \code{\link{polygon}}.}
}
\description{
Plot time series as stacked polygons with color fill and/or shading.
}
\examples{

## test data
set.seed(135)
dat <- data.frame(time = seq(7, 365, 14),
                  y1 = runif(26, min=1, max=10),
                  y2 = runif(26, min=0, max=15),
                  y3 = runif(26, min=0, max=2)
)


##' example palette
palette <-c("grey", "brown", "green", "yellow", "yellowgreen",
            "darkgreen", "blue", "black", "red", "grey", "red", "orange",
            "pink", "cyan"
)


## usage
stackedpoly(dat, c("y1", "y2", "y3"), col=palette)

## polygons with border
stackedpoly(dat, c("y1", "y2", "y3"), col=palette, border="black")

## optional plotting parameters
stackedpoly(dat, c("y1", "y2", "y3"), col=palette, ylim=c(0,50), main="Phytoplankton")

## changed order
stackedpoly(dat, c("y3", "y2", "y1"), col=palette, main="Phytoplankton")

## hatched area
stackedpoly(dat, c("y3", "y2", "y1"), density=c(20, 20, 40),
            angle=c(0, 90, 45), ylim=c(0,50), border=palette, main="Phytoplankton")

}
\seealso{
\code{\link{plot.default}}, \code{\link{polygon}}
}

